refactor(app): split local-resources/instruments/hooks
